Refactor git services into source control layers
- Move GitManager and GitHubCli contracts/layers to their new modules
- Update server wiring and tests for the relocated pluggable git setup
