+++ apps/server/src/sourceControl/SourceControlProviderRegistry.test.ts
@@ -1,7 +1,7 @@
 import { assert, it } from "@effect/vitest";
 import { Effect, Layer } from "effect";
 
-import { GitHubCli } from "../git/Services/GitHubCli.ts";
+import { GitHubCli } from "./GitHubCli.ts";
 import * as GitVcsDriver from "../vcs/GitVcsDriver.ts";
 import * as SourceControlProviderRegistry from "./SourceControlProviderRegistry.ts";
